Refactor setReady method in ComingController.php
Update the date format in the setReady method of ComingController.php to use the 'Y/m/d H:i:s' format instead of 'd/m/Y H:i:s'. Additionally, modify the logic to handle the case when the record does not exist in the 'dbo.InfoComming' table, either updating or inserting the 'ready' and 'IDRuchuMagazynowego' values accordingly.